feat(facts-db-adapter): add AbstractAdapter class
This commit adds the AbstractAdapter class to the facts-db-adapter package. The AbstractAdapter class is an abstract class that implements the IAdapter interface. It provides common functionality for adapters that interact with different tables in the facts-db package. The class includes methods for adding, deleting, and retrieving items from the specified table, as well as an observable for tracking updates. This class will serve as a base for specific adapters in the future.